Quick Fixes Based on Cleanup
