+++ client/src/utils/dateUtil.js
@@ -1,0 +1,22 @@
+const dateToYMD = (date) => {
+  var strArray = [
+    "Jan",
+    "Feb",
+    "Mar",
+    "Apr",
+    "May",
+    "Jun",
+    "Jul",
+    "Aug",
+    "Sep",
+    "Oct",
+    "Nov",
+    "Dec",
+  ];
+  var d = date.getDate();
+  var m = strArray[date.getMonth()];
+  var y = date.getFullYear();
+  return "" + (d <= 9 ? "" + d : d) + " " + m + " " + y;
+};
+
+module.exports = { dateToYMD };
